+++ next.config.ts
@@ -10,7 +10,8 @@ const nextConfig: NextConfig = {
     })
 
     return config
-  }
+  },
+  reactStrictMode: false,
 };
 
 export default nextConfig;
